refactor: move input event handling into its own file
